CatValidator and BirdValidator are created, Validataion in CatController and BirdController implemented.

+++ API/Controllers/CatsController/CatsController.cs
@@ -4,11 +4,11 @@ using Application.Commands.Cats.UpdateCat;
 using Application.Dtos;
 using Application.Queries.Cats.GetAllCats;
 using Application.Queries.Cats.GetCatById;
+using Application.Validators.Cat;
+using Application.Validators.GuidValidator;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace API.Controllers.CatsController
 {
@@ -19,17 +19,29 @@ namespace API.Controllers.CatsController
     {
         //Create mediator for comunication with DB
         private readonly IMediator _mediator;
+        private readonly CatValidator _catValidator;
+        private readonly GuidValidator _guidValidator;
 
-        public CatsController(IMediator mediator)
+        public CatsController(IMediator mediator, CatValidator catValidator, GuidValidator guidValidator)
         {
             _mediator = mediator;
+            _catValidator = catValidator;
+            _guidValidator = guidValidator;
         }
 
         [HttpGet]
         [Route("getAllCats"), AllowAnonymous]
         public async Task<IActionResult> GetAllCats()
         {
-            return Ok(await _mediator.Send(new GetAllCatsQuery()));
+            try
+            {
+                return Ok(await _mediator.Send(new GetAllCatsQuery()));
+            }
+            catch (Exception ex)
+            {
+
+                throw new Exception(ex.Message);
+            }
         }
 
         // GET api
@@ -37,7 +49,22 @@ namespace API.Controllers.CatsController
         [Route("getCatById/{catId}"), AllowAnonymous]
         public async Task<IActionResult> GetCatById(Guid catId)
         {
-            return Ok(await _mediator.Send(new GetCatByIdQuery(catId)));
+            // Validate Guid id
+            var validatedGuidId = _guidValidator.Validate(catId);
+            //Error Handling
+            if (!validatedGuidId.IsValid)
+            {
+                return BadRequest(validatedGuidId.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+            //Try Catch
+            try
+            {
+                return Ok(await _mediator.Send(new GetCatByIdQuery(catId)));
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
         }
 
         // Create a new cat 
@@ -45,7 +72,22 @@ namespace API.Controllers.CatsController
         [Route("addNewCat"), Authorize(Roles = "admin")]
         public async Task<IActionResult> AddCat([FromBody] CatDto newCat)
         {
-            return Ok(await _mediator.Send(new AddCatCommand(newCat)));
+            // Validate Cat
+            var validatedCat = _catValidator.Validate(newCat);
+            //Error Handling
+            if (!validatedCat.IsValid)
+            {
+                return BadRequest(validatedCat.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+            //Try Catch
+            try
+            {
+                return Ok(await _mediator.Send(new AddCatCommand(newCat)));
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
         }
 
         // Update info of a specific cat
@@ -53,7 +95,27 @@ namespace API.Controllers.CatsController
         [Route("updateCatInfo/{updatedCatId}"), Authorize(Roles = "admin")]
         public async Task<IActionResult> UpdateCat([FromBody] CatDto updatedCat, Guid updatedCatId)
         {
-            return Ok(await _mediator.Send(new UpdateCatInfoByIdCommand(updatedCat, updatedCatId)));
+            // Validate updatedCat, Guid updatedCatId
+            var validatedUpdatedCat = _catValidator.Validate(updatedCat);
+            var validatedUpdatedCatId = _guidValidator.Validate(updatedCatId);
+            //Error Handling
+            if (!validatedUpdatedCat.IsValid)
+            {
+                return BadRequest(validatedUpdatedCat.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+            if (!validatedUpdatedCatId.IsValid)
+            {
+                return BadRequest(validatedUpdatedCatId.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+            //Try Catch
+            try
+            {
+                return Ok(await _mediator.Send(new UpdateCatInfoByIdCommand(updatedCat, updatedCatId)));
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
         }
 
         // Delete a specific cat
@@ -61,7 +123,22 @@ namespace API.Controllers.CatsController
         [Route("deleteCat/{catToDeleteId}"), Authorize(Roles = "admin")]
         public async Task<IActionResult> DeleteCat(Guid catToDeleteId)
         {
-            return Ok(await _mediator.Send(new DeleteCatByIdCommand(catToDeleteId)));
+            // Validate Guid catToDeleteId
+            var validateCatToDeleteId = _guidValidator.Validate(catToDeleteId);
+            //Error Handling
+            if (!validateCatToDeleteId.IsValid)
+            {
+                return BadRequest(validateCatToDeleteId.Errors.ConvertAll(errors => errors.ErrorMessage));
+            }
+            //Try Catch
+            try
+            {
+                return Ok(await _mediator.Send(new DeleteCatByIdCommand(catToDeleteId)));
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
         }
     }
 }
